requirements/runtime.txt
django==1.11
python-docx==0.8.6
reportlab==3.4.0